restructure renderer package
