+++ packages/core/src/config/image-format.ts
@@ -1,4 +1,4 @@
-import {PixelFormat} from '.';
+import {PixelFormat} from './pixel-format';
 
 const validOptions = ['png', 'jpeg', 'none'] as const;
