+++ packages/media-utils/src/use-audio-metadata.ts
@@ -1,6 +1,6 @@
 import {useCallback, useEffect, useRef, useState} from 'react';
 import {continueRender, delayRender} from 'remotion';
-import {getAudioData} from '.';
+import {getAudioData} from './get-audio-data';
 import {AudioData} from './types';
 
 export const useAudioData = (src: string): AudioData | null => {
