+++ packages/renderer/src/open-browser.ts
@@ -1,0 +1,34 @@
+import puppeteer from 'puppeteer-core';
+import {Browser, Internals} from 'remotion';
+import {
+	ensureLocalBrowser,
+	getLocalBrowserExecutable,
+} from './get-local-browser-executable';
+
+export const openBrowser = async (
+	browser: Browser,
+	options?: {
+		shouldDumpIo?: boolean;
+	}
+): Promise<puppeteer.Browser> => {
+	if (browser === 'firefox' && !Internals.FEATURE_FLAG_FIREFOX_SUPPORT) {
+		throw new TypeError(
+			'Firefox supported is not yet turned on. Stay tuned for the future.'
+		);
+	}
+	await ensureLocalBrowser(browser);
+
+	const executablePath = await getLocalBrowserExecutable(browser);
+	const browserInstance = await puppeteer.launch({
+		executablePath,
+		product: browser,
+		dumpio: options?.shouldDumpIo ?? false,
+		args: [
+			'--no-sandbox',
+			'--disable-setuid-sandbox',
+			'--disable-dev-shm-usage',
+			process.platform === 'linux' ? '--single-process' : null,
+		].filter(Boolean) as string[],
+	});
+	return browserInstance;
+};
